don't reprocess data if output file's modification date is later/greater than the most recent input file's modification date

+++ 3-aggregate-all.py
@@ -1,6 +1,7 @@
 import importlib
 
 import argparse
+import glob
 import os
 from os import listdir
 from os.path import isfile, join
@@ -18,10 +19,17 @@ class Launcher():
         thisDir = os.path.dirname(os.path.abspath(__file__))
         idir = opts['aggregate_dir']
         for prefix in ['RC_', 'RS_']:
+
+            output_file = join(opts['aggregate_all_dir'], prefix+'aggregate_all.csv')
+            files = list(filter(isfile, glob.glob(idir + f"{prefix}*.csv")))
+            files.sort(key=lambda x: os.path.getmtime(x), reverse=True)
+            mostRecentInputFileModificationTime = os.path.getmtime(files[0])
+            if not(opts['force']) and isfile(output_file) and os.path.getmtime(output_file) > mostRecentInputFileModificationTime:
+                log('skipping '+prefix)
+                continue
             log('3-aggregate-all '+prefix)
             aa = AggregateAll(idir, opts['aggregate_all_min_rows'], opts['aggregate_n_rows'])
             aa.process(prefix)
-            output_file = join(opts['aggregate_all_dir'], prefix+'aggregate_all.csv')
             aa.write_csv(output_file)
         log('finished')
